Introduce BulkApiErrorContent composite; use in existing bulk pages; add API error handling to MultipleUserEnrollmentWorkflow
